@@ -5,7 +5,7 @@ author: fengqiangboy <fengqiangboy@gmail.com>
 homepage: https://github.com/fengqiangboy/flutter-nordic-dfu
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
